refactor(channel): create base class
